requirements.txt
asgiref==3.2.10
astroid==2.4.2
click==7.1.2
colorama==0.4.4
Django==3.1.1
djangorestframework==3.12.2
fastapi==0.61.1
future==0.18.2
importlib-metadata==1.7.0
isort==5.6.4
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
livereload==2.6.3
lunr==0.5.8
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
nltk==3.5
pydantic==1.6.1
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
starlette==0.13.6
toml==0.10.2
tornado==6.0.4
tqdm==4.48.2
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
